activitylog/api/requests.go
// Copyright (c) Abstract Machines
// SPDX-License-Identifier: Apache-2.0

package api

import (
	"github.com/absmach/magistrala/activitylog"
	"github.com/absmach/magistrala/internal/api"
	"github.com/absmach/magistrala/internal/apiutil"
)

type retrieveActivitiesReq struct {
	token string
	page  activitylog.Page
}

func (req retrieveActivitiesReq) validate() error {
	if req.token == "" {
		return apiutil.ErrBearerToken
	}
	if req.page.Limit > api.DefLimit {
		return apiutil.ErrLimitSize
	}
	if req.page.Direction != "" && req.page.Direction != api.AscDir && req.page.Direction != api.DescDir {
		return apiutil.ErrInvalidDirection
	}
	if req.page.EntityID != "" && req.page.EntityType == activitylog.EmptyEntity {
		return apiutil.ErrMissingEntityType
	}
	if req.page.EntityID == "" && req.page.EntityType != activitylog.EmptyEntity {
		return apiutil.ErrMissingID
	}

	return nil
}
